pyproject.toml
[project]
name = "dslab-rag-e2e"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.116.1",
    "fastembed>=0.7.3",
    "langchain>=0.3.27",
    "langchain-mistralai>=0.2.11",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "matplotlib>=3.10.6",
    "mistralai>=1.9.10",
    "numpy>=2.2.6",
    "openai>=1.106.1",
    "pandas>=2.3.2",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "qdrant-client>=1.15.1",
    "requests>=2.32.5",
    "uvicorn>=0.35.0",
]